Title: 세 수의 합과 평균 구하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240729/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.cpp
@@ -5,6 +5,6 @@ int main() {
     int a, b, c;
     cin >> a >>b >>c;
     cout << a + b + c << "\n";
-    cout << (double)(a + b + c) / 3;
+    cout << (a + b + c) / 3;
     return 0;
 }
